Split plugin framework into its own package
