+++ pkg/plugin/framework/backup_item_action_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package plugin
+package framework
 
 import (
 	"encoding/json"
