+++ pkg/plugin/framework/doc.go
@@ -1,0 +1,3 @@
+// Package framework is the common package that any plugin client
+// will need to import, for example, both plugin authors and Velero core.
+package framework
