+++ pkg/plugin/framework/plugin_base_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 the Velero contributors.
+Copyright 2018, 2019 the Velero contributors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -13,19 +13,28 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-
-package plugin
+package framework
 
 import (
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+
+	"github.com/heptio/velero/pkg/util/test"
 )
 
-// handlePanic is a panic handler for the server half of velero plugins.
-func handlePanic(p interface{}) error {
-	if p == nil {
-		return nil
-	}
+func TestClientLogger(t *testing.T) {
+	base := &pluginBase{}
+	logger := test.NewLogger()
+	f := ClientLogger(logger)
+	f(base)
+	assert.Equal(t, logger, base.clientLogger)
+}
 
-	return status.Errorf(codes.Aborted, "plugin panicked: %v", p)
+func TestServerLogger(t *testing.T) {
+	base := &pluginBase{}
+	logger := test.NewLogger()
+	f := serverLogger(logger)
+	f(base)
+	assert.Equal(t, newServerMux(logger), base.serverMux)
 }
