+++ pkg/plugin/process.go
@@ -19,6 +19,8 @@ import (
 	plugin "github.com/hashicorp/go-plugin"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
+
+	"github.com/heptio/velero/pkg/plugin/framework"
 )
 
 type ProcessFactory interface {
@@ -75,13 +77,13 @@ func (r *process) dispense(key kindAndName) (interface{}, error) {
 	}
 
 	// Currently all plugins except for PluginLister dispense clientDispenser instances.
-	if clientDispenser, ok := dispensed.(ClientDispenser); ok {
+	if clientDispenser, ok := dispensed.(framework.ClientDispenser); ok {
 		if key.name == "" {
 			return nil, errors.Errorf("%s plugin requested but name is missing", key.kind.String())
 		}
 		// Get the instance that implements our plugin interface (e.g. ObjectStore) that is a gRPC-based
 		// client
-		dispensed = clientDispenser.clientFor(key.name)
+		dispensed = clientDispenser.ClientFor(key.name)
 	}
 
 	return dispensed, nil
